fix: use postCreateCommand instead of postStartCommand
In all cases so far, the usage was slightly wrong: for example, installing am NPM tool must only happen once upon creation. The same is true for the restore of global .NET tools or NuGet packages, which doesn't have to happen on every container start. For Jekyll, this could even lead to blocked ports, i.e. jekyll serve must only happen once but not every time when starting the container.